feat(notification): initial implementation
The implementation tries to translate what homeassistant provides to the
freedesktop spec.

freedesktop:
  - https://people.gnome.org/~mccann/docs/notification-spec/notification-spec-latest.html#basic-design

homeassistant:
- https://companion.home-assistant.io/docs/notifications/notifications-basic/

Supports:
  - Urgency
  - Timeout
  - Actions
    - Action callbacks [WIP]: hooks are in place but not implemented
  - Events
    - Dismissal
    - Event callbacks [WIP]: hooks are in place but not implemented

Implementation:

A web server listens for Home Assistant HTTP POST in an specific route
"/notify" which invokes a handler that translates the data from the
format Home Assistant uses to what the freedesktop specifies. This data
is then sent over dbus to the desktop, then a notification server will
display it.

When a notification is dismissed, or an action is selected a message is
sent over dbus, we subscribe to listen to those messges with a callback.

Next step is to handle some of the actions like opening a url, or
launching a program and also emit that an action was selected back to
Home Assistant.