refactor(all): split logic, improve clarity
API was full of stuff, moving things to a sensor manager and
registration of the companion, to the companion objects make more sense.

In the sensor manager side, in the future support is planned to update
sensors on dbus signals which would need some extra logic that would've
filled API of unnecessary things.

In the companion side, well it makes sense to move registering of the
companion to it's class, also decided to just return the data and send
it to the api, to avoid mutating api from another class.

Also added more comments explaining the application behavior.