feat(sensor signals): callback to update sensors
Now sensors are able to register a set of pre defined signals, at the
moment dbus only, that when fired the sensor methods will be called,
making it able to update it's state, and after an update_sensor will be
fired.

This is done by using a partial function that calls the sensor provided
callback from the sensor manager.