feat(notification): allow running commands
Commands are defined in the config, with format "command_{name}" it
defines a "name" which is a description of the command, and the actual
command which is a list of arguments.

The arguments are passed to asyncio.create_subprocess_exec, stdout and
stderr are piped to /dev/null, no error catching is done.

Also simplify the structure of config.json moving url_program under
notification service and also the commands, since all is coming through
this medium.